passes: Refactor with langbase.py and parpragma.py
