+++ devito/passes/iet/languages/__init__.py
@@ -1,3 +1,2 @@
-from .basic import is_on_device  # noqa
 from .openmp import *  # noqa
 from .openacc import *  # noqa
